Refactor resolv.conf code
Move handling of resolv.conf to a separate file.
