+++ CMakeLists.txt
@@ -16,6 +16,7 @@ LIST(APPEND neat_SOURCES
     neat_addr.c
     neat_he.c
     neat_resolver.c
+    neat_resolver_conf.c
     neat_property_helpers.c
     neat_log.c
     )
